+++ Infrastructure.Repository/DistanceMatrixRepositoryCollection/DistanceMatrixRepository.cs
@@ -6,29 +6,29 @@ namespace Infrastructure.Repository.DistanceMatrixRepositoryCollection
 {
     public class DistanceMatrixRepository : IDistanceMatrixRepository
     {
-        private readonly Dictionary<int, Dictionary<int, DistanceInfo>> _distanceMatrix;
+        private readonly Dictionary<long, Dictionary<long, DistanceInfo>> _distanceMatrix;
 
         public DistanceMatrixRepository()
         {
-            this._distanceMatrix = new Dictionary<int, Dictionary<int, DistanceInfo>>();
+            this._distanceMatrix = new Dictionary<long, Dictionary<long, DistanceInfo>>();
         }
         
-        public void Add(int originCustomer, int destinationCustomer, DistanceInfo distanceInfo)
+        public void Add(long originCustomer, long destinationCustomer, DistanceInfo distanceInfo)
         {
             
             if(!this._distanceMatrix.ContainsKey(originCustomer))
-                _distanceMatrix.Add(originCustomer, new Dictionary<int, DistanceInfo>());
+                _distanceMatrix.Add(originCustomer, new Dictionary<long, DistanceInfo>());
             
             _distanceMatrix[originCustomer].Add(destinationCustomer, distanceInfo);
 
         }
 
-        public float GetDistance(int originCustomer, int destinationCustomer)
+        public float GetDistance(long originCustomer, long destinationCustomer)
         {
             return _distanceMatrix[originCustomer][destinationCustomer].Distance;
         }
 
-        public float GetTime(int originCustomer, int destinationCustomer)
+        public float GetTime(long originCustomer, long destinationCustomer)
         {
             return _distanceMatrix[originCustomer][destinationCustomer].Time;
         }
